Caf Field: Comments etc. regarding optionvalues
Fields with option values implicitly have an index based
seUiValue/uiValue interface. This is commented better. Fields of AppEnum
specialization was not quite in line with the rest of the optionValue
system. That is commented and some ifdeffed code has been added to
highlight what could be done when we dare.